Back-End/requirements.txt
# FastAPI Backend Dependencies for Positivity Push

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# External Service Integrations
stripe==7.8.0
supabase==2.1.0
openai==1.3.0

# Background Processing
celery[redis]==5.3.4
redis==5.0.1

# Core Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
requests==2.31.0
httpx==0.25.2

# AI Memory & Context
mem0ai==0.0.8

# Email Service
sendgrid==6.11.0

# Additional Utilities
qrcode[pil]==7.4.2
python-dateutil==2.8.2
typing-extensions==4.8.0